CORE-4190: long

+++ sharedbuffers/mapped_struct.py
@@ -438,7 +438,7 @@ def shared_id(obj):
             return obj | LONG_MASK
         elif isinstance(obj, long):
             # Real longs must make room for the flag bits
-            return (obj & 0xFFFFFFFFFFFFFFFF) | ((obj >> 64) << 68) | LONG_MASK
+            return (obj & long(0xFFFFFFFFFFFFFFFF)) | ((obj >> 64) << 68) | LONG_MASK
 
         # For other keys, use the object itself as key if hashable
         try:
@@ -456,10 +456,10 @@ class WRAPPED:
     pass
 
 
-WRAP_MASK = cython.declare(object, 1 << 64)
-LONG_MASK = cython.declare(object, 2 << 64)
-PROXY_MASK = cython.declare(object, 4 << 64)
-FLAG_MASK = cython.declare(object, 0xF << 64)
+WRAP_MASK = cython.declare(object, long(1) << 64)
+LONG_MASK = cython.declare(object, long(2) << 64)
+PROXY_MASK = cython.declare(object, long(4) << 64)
+FLAG_MASK = cython.declare(object, long(0xF) << 64)
 
 
 @cython.ccall
@@ -742,15 +742,15 @@ class mapped_tuple(tuple):
                         # inline unsigned ints
                         dtype = b'I'
                         buf[offs] = ord(dtype)
-                    elif (cython.cast(cython.longlong, -0x8000000000000000) <= iminval
-                            and imaxval <= cython.cast(cython.longlong, 0x7FFFFFFFFFFFFFFF)):
+                    elif (cython.cast(cython.longlong, long(-0x8000000000000000)) <= iminval
+                            and imaxval <= cython.cast(cython.longlong, long(0x7FFFFFFFFFFFFFFF))):
                         # inline signed int64 list
                         buf[offs] = ord('q')
                         dtype = b'l'
                     else:
                         raise OverflowError
                 except OverflowError:
-                    if 0 <= minval and maxval <= 0xFFFFFFFFFFFFFFFF:
+                    if 0 <= minval and maxval <= long(0xFFFFFFFFFFFFFFFF):
                         # inline unsigned int64 list
                         buf[offs] = ord('Q')
                         dtype = b'L'
