+++ tests/mapped_struct.py
@@ -1771,7 +1771,7 @@ class ProxiedFrozensetPackingTest(unittest.TestCase, CommonCollectionPackingTest
         self.assertTrue(big >= small)
         self.assertFalse(big > big)
 
-        big = self.pack([1, 2, 3])
+        big = self.pack([long(1), long(2), long(3)])
         self.assertTrue(small < big)
         self.assertTrue(small <= big)
 
@@ -2241,11 +2241,11 @@ class StableHashTest(unittest.TestCase):
             (20913041029, 20913041029),
             (66210231110, 66210231110),
             (752, 752),
-            (-1, 18446744073709551615),
+            (-1, long(18446744073709551615)),
             (1 << 100, 1),
             (-(1 << 100), 1),
             (1, 1),
-            (-1, 18446744073709551615),
+            (-1, long(18446744073709551615)),
         )
         self.assertHashesOK(values_hashes)
 
@@ -2256,28 +2256,28 @@ class StableHashTest(unittest.TestCase):
             (107203., 107203),
             (91024215., 91024215),
             (13328914., 13328914),
-            (-1., 18446744073709551615),
-            (1e+50, 150463276902340),
-            (1e-50, 263280728297184),
-            (2.001, 140807857099441),
-            (-2.001, 18446603265852452175),
-            (float('inf'), 281474976710655),
-            (float('-inf'), 18446462598732840961),
-            (float('nan'), 562949953421310),
+            (-1., long(18446744073709551615)),
+            (1e+50, long(150463276902340)),
+            (1e-50, long(263280728297184)),
+            (2.001, long(140807857099441)),
+            (-2.001, long(18446603265852452175)),
+            (float('inf'), long(281474976710655)),
+            (float('-inf'), long(18446462598732840961)),
+            (float('nan'), long(562949953421310)),
         )
         self.assertHashesOK(values_hashes)
 
     def testHashStringsFixed(self):
         values_hashes = (
-            ("abcdef", 18053520794346263629),
-            ("123456789", 10139926970967174787),
-            ("!@#%&$", 15648343848775299486),
+            ("abcdef", long(18053520794346263629)),
+            ("123456789", long(10139926970967174787)),
+            ("!@#%&$", long(15648343848775299486)),
         )
         self.assertHashesOK(values_hashes)
 
     def testHashSequenceFixed(self):
         values_hashes = (
-            ((1, "abc", -2.3), 2059039662577021167),
+            ((1, "abc", -2.3), long(2059039662577021167)),
             # FIXME this doesnt work, the frozensets have differente order in py3 vs py2
             # (frozenset([1.2, 4.5, 0.12]), 4015877951310865576),
         )
@@ -2463,8 +2463,8 @@ class WriteableMappingTest(unittest.TestCase):
         proxy.hx = -1000
         proxy.Ix = 100000
         proxy.ix = -100000
-        proxy.Qx = 10000000000
-        proxy.qx = -10000000000
+        proxy.Qx = long(10000000000)
+        proxy.qx = long(-10000000000)
         proxy.dx = 1.5
         proxy.fx = -1.5
         self.assertAlmostEqual(0.0, proxy.Bx + proxy.bx + proxy.Hx + proxy.hx +
